# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: SafeSphere API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sos/trigger:
    post:
      operationId: triggerSos
      tags: [sos]
      summary: Trigger SOS — creates an incident, sends alerts to contacts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerSosBody"
      responses:
        "201":
          description: Incident created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
  /sos/{id}/cancel:
    post:
      operationId: cancelIncident
      tags: [sos]
      summary: Cancel an active incident (false alarm)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Incident cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
  /sos/{id}/resolve:
    post:
      operationId: resolveIncident
      tags: [sos]
      summary: Mark incident as resolved (safe)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
  /sos/{id}/escalate-nearby:
    post:
      operationId: escalateNearby
      tags: [sos]
      summary: Escalate to nearby SafeSphere users when no response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Escalation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscalationResult"
  /sos/{id}/motion:
    post:
      operationId: reportMotion
      tags: [sos]
      summary: Report device motion sample for AI risk scoring
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MotionSampleBody"
      responses:
        "200":
          description: Updated risk
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskUpdate"
  /sos/{id}/responders:
    get:
      operationId: getIncidentResponders
      tags: [responders]
      summary: List responders engaged on this incident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Responders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IncidentResponder"
...
components:
  schemas:
...
    TriggerSosBody:
      type: object
      required: [trigger, lat, lng]
      properties:
        trigger:
          type: string
          description: button | voice | lock-screen | shake
        lat: { type: number }
        lng: { type: number }
        accuracy:
          type: ["number", "null"]
        discreet:
          type: ["boolean", "null"]
        message:
          type: ["string", "null"]
    MotionSampleBody:
      type: object
      required: [speed]
      properties:
        speed:
          type: number
          description: m/s
        accel:
          type: ["number", "null"]
          description: m/s² peak
        repeated:
          type: ["boolean", "null"]
          description: User pressed SOS multiple times
    RiskUpdate:
      type: object
      required: [incidentId, riskScore, riskLevel, escalationSeconds]
      properties:
        incidentId: { type: integer }
        riskScore: { type: integer }
        riskLevel: { type: string }
        escalationSeconds: { type: integer }
    Incident:
      type: object
      required:
        [
          id,
          status,
          trigger,
          startLat,
          startLng,
          alertsSent,
          escalated,
          discreet,
          riskScore,
          riskLevel,
          createdAt,
        ]
      properties:
        id: { type: integer }
        status: { type: string }
        trigger: { type: string }
        startLat: { type: number }
        startLng: { type: number }
        alertsSent: { type: integer }
        escalated: { type: boolean }
        discreet: { type: boolean }
        riskScore: { type: integer }
        riskLevel: { type: string }
        motionMaxSpeed:
          type: ["number", "null"]
        triggerCount24h:
          type: ["integer", "null"]
        message:
          type: ["string", "null"]
        resolvedAt:
          type: ["string", "null"]
          format: date-time
        createdAt: { type: string, format: date-time }
...
    EscalationResult:
      type: object
      required: [incidentId, nearbyAlerted, contactsNotified]
      properties:
        incidentId: { type: integer }
        nearbyAlerted: { type: integer }
        contactsNotified: { type: integer }
...
    IncidentResponder:
      type: object
      required:
        [id, incidentId, alias, status, distanceKm, createdAt]
      properties:
        id: { type: integer }
        incidentId: { type: integer }
        alias: { type: string }
        status: { type: string }
        distanceKm: { type: number }
        etaMinutes:
          type: ["integer", "null"]
        createdAt: { type: string, format: date-time }
        respondedAt:
          type: ["string", "null"]
          format: date-time
...